refactored graph component
