+++ dashboard/src/components/SmellGraph.js
@@ -1,8 +1,10 @@
 import React, {Component} from "react"
 import { Bar } from "react-chartjs-2"
 import {Col, Dropdown, DropdownButton, Row} from "react-bootstrap"
-import Sketch from "@uiw/react-color-sketch"
-import {CirclePicker} from "react-color"
+import PrimaryPicker from "./PrimaryPicker";
+import {ResetPrimary, ResetSecondary} from "./ResetButtons";
+import SecondaryPicker from "./SecondaryPicker";
+import {FontDropdown} from "./FontDropdown";
 
 export default class SmellGraph extends Component {
     constructor(props) {
@@ -13,6 +15,32 @@ export default class SmellGraph extends Component {
             fontSize: localStorage.getItem('fontSize') != null ? localStorage.getItem('fontSize') : 16,
             colors: localStorage.getItem('colors') !== null ? localStorage.getItem('colors').split(',') : (localStorage.getItem('hex') !== null ? localStorage.getItem('hex') : ['#009688'])
         }
+        this.changePrimaryColor = this.changePrimaryColor.bind(this)
+        this.changeSecondaryColor = this.changeSecondaryColor.bind(this)
+        this.changeFontSize = this.changeFontSize.bind(this)
+    }
+
+    changePrimaryColor(colors) {
+        this.setState({
+            hex: colors[0],
+            colors: colors
+        })
+        localStorage.setItem('hex', colors[0])
+        localStorage.setItem('colors', colors)
+    }
+
+    changeSecondaryColor(colors) {
+        this.setState({
+            colors: colors
+        })
+        localStorage.setItem('colors', colors)
+    }
+
+    changeFontSize(size) {
+        this.setState({
+            fontSize: size
+        })
+        localStorage.setItem('fontSize', size)
     }
 
     render() {
@@ -59,37 +87,10 @@ export default class SmellGraph extends Component {
                             </Dropdown.Toggle>
                             <Dropdown.Menu>
                                 <Dropdown.ItemText>Primary Color</Dropdown.ItemText>
-                                <Dropdown.ItemText>
-                                    <Sketch
-                                        color={hex}
-                                        presetColors={['#ff0000', '#ffa500', '#ffff00', '#00ff00',
-                                                       '#009688', '#0000ff', '#9300ff', '#e01fe0']}
-                                        onChange={(color) => {
-                                            colors[0] = color.hex
-                                            this.setState({
-                                                hex: color.hex,
-                                                colors: colors
-                                            })
-                                            localStorage.setItem('hex', color.hex)
-                                            localStorage.setItem('colors', colors)
-                                        }}
-                                    />
-                                </Dropdown.ItemText>
-                                <Dropdown.Item
-                                    className='reset-button'
-                                    onClick={() => {
-                                        colors[0] = defaultHex
-                                        this.setState({
-                                            hex: defaultHex,
-                                            colors: colors
-                                        })
-                                        localStorage.setItem('hex', defaultHex)
-                                        localStorage.setItem('colors', defaultHex)
-                                    }}
-                                >
-                                    Reset
-                                </Dropdown.Item>
+                                    <PrimaryPicker hex={hex} colors={colors} changePrimaryColor={this.changePrimaryColor}/>
+                                    <ResetPrimary defaultHex={defaultHex} colors={colors} changePrimaryColor={this.changePrimaryColor}/>
                                 <Dropdown.Divider />
+
                                 <DropdownButton
                                     id='circle-picker'
                                     title="Add Color"
@@ -97,64 +98,17 @@ export default class SmellGraph extends Component {
                                     drop='end'
                                     onSelect
                                 >
-                                    <Dropdown.ItemText>
-                                        <CirclePicker
-                                            width={170}
-                                            colors={['#990000', '#ff0000', '#ff6666', '#ffcccc',
-                                                     '#996300', '#ffa500', '#ffc966', '#ffedcc',
-                                                     '#999900', '#ffff00', '#ffff66', '#ffffcc',
-                                                     '#009900', '#00ff00', '#66ff66', '#ccffcc',
-                                                     '#00998b', '#00ffe7', '#80FFF3', '#ccfffa',
-                                                     '#000099', '#0000ff', '#6666ff', '#ccccff',
-                                                     '#580099', '#9300ff', '#be66ff', '#e9ccff',
-                                                     '#871287', '#e01fe0', '#ed78ed', '#f9d2f9',]}
-                                            onChangeComplete={(color) => {
-                                                colors.push(color.hex)
-                                                this.setState({
-                                                    colors: colors
-                                                })
-                                                localStorage.setItem('colors', colors)
-                                            }}
-                                        />
-                                    </Dropdown.ItemText>
-                                    <Dropdown.Item
-                                        className='reset-button'
-                                        onClick={() => {
-                                            let colors = [hex]
-                                            this.setState({
-                                                colors: colors
-                                            })
-                                            localStorage.setItem('colors', colors.toString())
-                                        }}
-                                    >
-                                        Reset
-                                    </Dropdown.Item>
+                                    <SecondaryPicker colors={colors} changeSecondaryColor={this.changeSecondaryColor}/>
+                                    <ResetSecondary hex={hex} colors={colors} changeSecondaryColor={this.changeSecondaryColor} />
                                 </DropdownButton>
                             </Dropdown.Menu>
                         </Dropdown>
                     </Col>
                     <Col>
-                        <DropdownButton
-                            title="Font"
-                            variant="secondary"
-                            id="fontDropdown"
-                            onSelect={(size) => {
-                                this.setState({
-                                    fontSize: size
-                                })
-                                localStorage.setItem('fontSize', size)
-                            }}
-                        >
-                            <Dropdown.Item eventKey="12">12</Dropdown.Item>
-                            <Dropdown.Item eventKey="16">16</Dropdown.Item>
-                            <Dropdown.Item eventKey="20">20</Dropdown.Item>
-                            <Dropdown.Item eventKey="24">24</Dropdown.Item>
-                            <Dropdown.Item eventKey="28">28</Dropdown.Item>
-                            <Dropdown.Item eventKey="32">32</Dropdown.Item>
-                            <Dropdown.Item eventKey="32">36</Dropdown.Item>
-                        </DropdownButton>
+                        <FontDropdown changeFontSize={this.changeFontSize} />
                     </Col>
                 </Row>
+
                 <Bar
                     className="padding-left-right"
                     data={
